Add Bearer token authentication and update frontend for file upload (#8)
* Add Bearer token authentication to file upload route and update frontend

* Integrate Minnisblad router and update upload route with token validation

* Update index.html to reflect branding changes and add new assistance button
