+++ .env.example
@@ -4,3 +4,4 @@ POSTGRES_SERVER=localhost
 POSTGRES_PORT=5432
 POSTGRES_DB=
 OPENAI_API_KEY=
+BEARER_TOKEN=
